requirements.txt
NumPy
Pandas
Matplotlib
Seaborn
scikit-learn
pyyaml